Fix support for OpenBSD
Fix the code that reads the cpu state for OpenBSD, and include support
for multiple CPUs. Patch by Landry Breuil.

+++ panel-plugin/os.h
@@ -8,8 +8,8 @@
 typedef struct
 {
 	guint load;
-	guint previous_used;
-	guint previous_total;
+	guint64 previous_used;
+	guint64 previous_total;
 } CpuData;
 
 guint detect_cpu_number();
